requirements.txt
pandas
pybit
python-dotenv
numpy
ta-lib
streamlit  # (opsional, jika ingin deploy dashboard juga)